Fix logic for legacy applet.
