+++ gtk_llm_chat/tk_llm_applet.py
@@ -19,7 +19,7 @@ from watchdog.events import FileSystemEventHandler
 
 # Importar el cliente D-Bus compartido
 sys.path.append(os.path.dirname(os.path.abspath(__file__)))
-from dbus_client import open_conversation_dbus
+from ipc_client import open_conversation_dbus
 
 # Importar historial de chat
 sys.path.append(os.path.dirname(os.path.abspath(__file__)))
